Nest exegesis's LspProxy under Lsp
`LspProxy` becomes `Lsp.Proxy`, with the companion reordered ahead of its class to
satisfy L3 (the donor had them the other way round).

The proposal's stated risk did not materialise: fourteen sites do `import Lsp.*`
alongside `import soundness.*`, so `Proxy` — a known homonym across austronesian
and vicarious — is now in scope unqualified at all of them, and none is ambiguous.

The move did need `aperture.*` added to `exegesis.Lsp.scala`: `upstream.session`
resolves through an extension method the donor file imported and the receiving
file did not. An extension method is exactly the kind of dependency that survives
a compile of the donor but not the move.
